+++ a1_driver/src/main.cpp
@@ -2,13 +2,13 @@
 // Created by sun on 2021/3/1.
 //
 
-#include "unitree_a1_ros2/A1_ros.h"
-#include "unitree_a1_ros2/A1_wrapper.h"
+#include "a1_driver/A1_ros.h"
+#include "a1_driver/A1_wrapper.h"
 
 int main(int argc, char *argv[]) {
     std::string node_name = "A1_node";
     if (argc < 3) {
-        std::cerr << "running ros2 run unitree_a1_ros2 a1_main --node-name A1_node" << std::endl;
+        std::cerr << "running ros2 run a1_driver a1_main --node-name A1_node" << std::endl;
     }
     for (int i = 1; i < argc; ++i) {
         std::string arg = argv[i];
